Extract business module: BusinessMixin + lazy tool registration
Follows the admin extraction pattern. Business is the largest module
(22 tools, ~2000 lines of methods) — biggest single win for the
--modules flag's memory semantics.

- book/business.py: BusinessMixin with all customer/vendor/billterm
  CRUD, invoice/bill creation and entries, posting, payment, delete
  paths, outstanding-balance reporting, and vendor spending report.
  Includes business-scoped helpers (_find_customer/vendor/invoice and
  _by_guid variants, _customer_to_dict and friends, _decimal_to_num_denom,
  _write_gncinvoice_slot, _write_gdate_slot, _calculate_lot_balance,
  _get_invoice_entries_and_total).
- tools/business.py: register(mcp, get_book) attaching all 22 business
  tool wrappers — imported only when --modules includes 'business'.
- book/__init__.py: add 'business' to _MIXIN_MAP.
- server.py: remove the 22 business tool wrappers.
- book/_legacy.py: remove the entire business block (-2017 lines).

MRO is now GnuCashBook -> LegacyMixin -> AdminMixin -> BusinessMixin
-> BaseGnuCashBook. The collision guard in build_book_class caught a
mid-refactor state where _find_customer existed in both LegacyMixin
and BusinessMixin, halting at import rather than silently shadowing.

Full suite: 602 passed. test_business.py's 49 tests run unchanged
against the new mixin (MRO preserves the method surface).

Live testing skipped for this PR: business features are niche and
the 49-test suite covers CRUD, entry management, posting, payment,
outstanding balances, vendor spending, and cross-currency flows.

At-import tool count: 49 (71 after admin PR, -22 business).
