+++ src/gnucash_mcp/book/__init__.py
@@ -28,6 +28,7 @@ from gnucash_mcp.book._legacy import LegacyMixin
 # As each module is peeled off the monolith, add an entry here.
 _MIXIN_MAP: dict[str, tuple[str, str]] = {
     "admin": (".admin", "AdminMixin"),
+    "business": (".business", "BusinessMixin"),
 }
