+++ src/gnucash_mcp/book/_legacy.py
@@ -4602,2021 +4602,4 @@ class LegacyMixin:
 
     # Account slot methods (get/set/delete) moved to book/admin.py as AdminMixin.
 
-    # ------------------------------------------------------------------
-    # Business: Customer / Vendor / Billterm
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def _find_customer(book, customer_id: str):
-        """Find a customer by their human-readable ID (e.g., '000001')."""
-        for c in book.customers:
-            if c.id == customer_id:
-                return c
-        return None
-
-    @staticmethod
-    def _find_vendor(book, vendor_id: str):
-        """Find a vendor by their human-readable ID (e.g., '000001')."""
-        for v in book.vendors:
-            if v.id == vendor_id:
-                return v
-        return None
-
-    @staticmethod
-    def _customer_to_dict(customer) -> dict:
-        """Convert a piecash Customer to a serializable dict."""
-        return {
-            "guid": customer.guid,
-            "id": customer.id,
-            "name": customer.name,
-            "currency": customer.currency.mnemonic if customer.currency else None,
-            "notes": customer.notes or "",
-            "active": bool(customer.active),
-            "address": {
-                "name": customer.addr_name or "",
-                "addr1": customer.addr_addr1 or "",
-                "addr2": customer.addr_addr2 or "",
-                "addr3": customer.addr_addr3 or "",
-                "addr4": customer.addr_addr4 or "",
-                "phone": customer.addr_phone or "",
-                "fax": customer.addr_fax or "",
-                "email": customer.addr_email or "",
-            },
-        }
-
-    @staticmethod
-    def _vendor_to_dict(vendor) -> dict:
-        """Convert a piecash Vendor to a serializable dict."""
-        return {
-            "guid": vendor.guid,
-            "id": vendor.id,
-            "name": vendor.name,
-            "currency": vendor.currency.mnemonic if vendor.currency else None,
-            "notes": vendor.notes or "",
-            "active": bool(vendor.active),
-            "address": {
-                "name": vendor.addr_name or "",
-                "addr1": vendor.addr_addr1 or "",
-                "addr2": vendor.addr_addr2 or "",
-                "addr3": vendor.addr_addr3 or "",
-                "addr4": vendor.addr_addr4 or "",
-                "phone": vendor.addr_phone or "",
-                "fax": vendor.addr_fax or "",
-                "email": vendor.addr_email or "",
-            },
-        }
-
-    @staticmethod
-    def _customer_to_compact_line(customer) -> str:
-        """One-line compact: 'id  name  currency'."""
-        return f"{customer.id}\t{customer.name}\t{customer.currency.mnemonic}"
-
-    @staticmethod
-    def _vendor_to_compact_line(vendor) -> str:
-        """One-line compact: 'id  name  currency'."""
-        return f"{vendor.id}\t{vendor.name}\t{vendor.currency.mnemonic}"
-
-    @staticmethod
-    def _billterm_to_dict(bt) -> dict:
-        """Convert a Billterm to a serializable dict."""
-        return {
-            "guid": bt.guid,
-            "name": bt.name,
-            "description": bt.description or "",
-            "type": bt.type,
-            "due_days": bt.duedays,
-            "discount_days": bt.discountdays if bt.discountdays else 0,
-            "discount": str(bt.discount) if bt.discount else "0",
-        }
-
-    @staticmethod
-    def _decimal_to_num_denom(value: Decimal) -> tuple[int, int]:
-        """Convert a Decimal to numerator/denominator pair.
-
-        Uses the decimal's exponent to determine appropriate denominator.
-        E.g., Decimal("25.50") -> (2550, 100)
-        """
-        sign, digits, exp = value.as_tuple()
-        if exp < 0:
-            denom = 10 ** (-exp)
-            num = int(value * denom)
-        else:
-            num = int(value)
-            denom = 1
-        return num, denom
-
-    @staticmethod
-    def _find_invoice(book, invoice_id: str, owner_type: int | None = None):
-        """Find an invoice/bill by human-readable ID.
-
-        Args:
-            book: piecash Book instance.
-            invoice_id: Human-readable ID (e.g., '000001').
-            owner_type: Filter by owner type (2=customer, 4=vendor).
-                        None returns first match of either type.
-        """
-        from piecash.business.invoice import Invoice
-        query = book.session.query(Invoice).filter(Invoice.id == invoice_id)
-        if owner_type is not None:
-            query = query.filter(Invoice.owner_type == owner_type)
-        return query.first()
-
-    @staticmethod
-    def _invoice_to_dict(invoice, entries=None) -> dict:
-        """Convert a piecash Invoice to a serializable dict.
-
-        Args:
-            invoice: piecash Invoice object.
-            entries: Optional list of entry dicts. If None, entries are not included.
-        """
-        is_posted = invoice.date_posted is not None
-        result = {
-            "guid": invoice.guid,
-            "id": invoice.id,
-            "type": "bill" if invoice.owner_type == 4 else "invoice",
-            "owner_type": invoice.owner_type,
-            "owner_guid": invoice.owner_guid,
-            "date_opened": str(invoice.date_opened.date()) if invoice.date_opened else None,
-            "date_posted": str(invoice.date_posted.date()) if invoice.date_posted else None,
-            "notes": invoice.notes or "",
-            "active": bool(invoice.active),
-            "currency": invoice.currency.mnemonic if invoice.currency else None,
-            "is_posted": is_posted,
-        }
-        if entries is not None:
-            result["entries"] = entries
-        return result
-
-    @staticmethod
-    def _invoice_to_compact_line(invoice) -> str:
-        """One-line compact: 'id  type  owner_id  date_opened  status'."""
-        inv_type = "BILL" if invoice.owner_type == 4 else "INV"
-        date_str = str(invoice.date_opened.date()) if invoice.date_opened else "n/a"
-        status = "posted" if invoice.date_posted else "open"
-        return f"{invoice.id}\t{inv_type}\t{date_str}\t{status}"
-
-    @staticmethod
-    def _entry_to_dict(entry_row, is_bill: bool = False) -> dict:
-        """Convert an entry row (from raw SQL) to a serializable dict.
-
-        Args:
-            entry_row: SQLAlchemy row from entries table.
-            is_bill: If True, read b_* columns; otherwise read i_* columns.
-        """
-        # Quantity
-        q_num = entry_row.quantity_num or 0
-        q_denom = entry_row.quantity_denom or 1
-        quantity = Decimal(q_num) / Decimal(q_denom)
-
-        if is_bill:
-            p_num = entry_row.b_price_num or 0
-            p_denom = entry_row.b_price_denom or 1
-            acct_guid = entry_row.b_acct
-        else:
-            p_num = entry_row.i_price_num or 0
-            p_denom = entry_row.i_price_denom or 1
-            acct_guid = entry_row.i_acct
-
-        price = Decimal(p_num) / Decimal(p_denom)
-        total = quantity * price
-
-        # Raw SQL returns date as string; handle both str and datetime
-        raw_date = entry_row.date
-        if raw_date is None:
-            date_str = None
-        elif isinstance(raw_date, str):
-            date_str = raw_date[:10]  # "YYYY-MM-DD..."
-        else:
-            date_str = str(raw_date.date())
-
-        return {
-            "guid": entry_row.guid,
-            "date": date_str,
-            "description": entry_row.description or "",
-            "account_guid": acct_guid or "",
-            "quantity": str(quantity),
-            "price": str(price),
-            "total": str(total),
-        }
-
-    def create_customer(
-        self,
-        name: str,
-        currency: str | None = None,
-        notes: str = "",
-        address: dict | None = None,
-    ) -> dict:
-        """Create a new customer.
-
-        Args:
-            name: Customer name.
-            currency: ISO currency code. Defaults to book's default currency.
-            notes: Optional notes.
-            address: Optional address dict with keys: name, addr1, addr2,
-                     addr3, addr4, phone, fax, email.
-
-        Returns:
-            Dict with guid, id, name, status.
-        """
-        from piecash.business.person import Customer, Address
-
-        with self.open(readonly=False) as book:
-            if currency:
-                currency_obj = None
-                for c in book.currencies:
-                    if c.mnemonic == currency:
-                        currency_obj = c
-                        break
-                if not currency_obj:
-                    raise ValueError(f"Currency not found: {currency}")
-            else:
-                currency_obj = self._require_default_currency(book)
-
-            addr = None
-            if address:
-                addr = Address(
-                    name=address.get("name", name),
-                    addr1=address.get("addr1", ""),
-                    addr2=address.get("addr2", ""),
-                    addr3=address.get("addr3", ""),
-                    addr4=address.get("addr4", ""),
-                    phone=address.get("phone", ""),
-                    fax=address.get("fax", ""),
-                    email=address.get("email", ""),
-                )
-
-            customer = Customer(
-                name=name,
-                currency=currency_obj,
-                notes=notes,
-                address=addr,
-                book=book,
-            )
-            book.save()
-
-            return {
-                "guid": customer.guid,
-                "id": customer.id,
-                "name": customer.name,
-                "status": "created",
-            }
-
-    def list_customers(
-        self,
-        active_only: bool = True,
-        compact: bool = True,
-    ) -> list[dict] | str:
-        """List all customers.
-
-        Args:
-            active_only: If True, only return active customers.
-            compact: If True, return compact one-line-per-customer string.
-
-        Returns:
-            Compact string or list of dicts.
-        """
-        with self.open() as book:
-            customers = sorted(book.customers, key=lambda c: c.name)
-            if active_only:
-                customers = [c for c in customers if c.active]
-
-            if compact:
-                lines = [self._customer_to_compact_line(c) for c in customers]
-                return "\n".join(lines)
-            else:
-                return [self._customer_to_dict(c) for c in customers]
-
-    def get_customer(self, customer_id: str) -> dict:
-        """Get customer details by ID.
-
-        Args:
-            customer_id: Human-readable customer ID (e.g., '000001').
-
-        Returns:
-            Dict with full customer details.
-
-        Raises:
-            ValueError: If customer not found.
-        """
-        with self.open() as book:
-            customer = self._find_customer(book, customer_id)
-            if not customer:
-                raise ValueError(f"Customer not found: {customer_id}")
-            return self._customer_to_dict(customer)
-
-    def create_vendor(
-        self,
-        name: str,
-        currency: str | None = None,
-        notes: str = "",
-        address: dict | None = None,
-    ) -> dict:
-        """Create a new vendor.
-
-        Args:
-            name: Vendor name.
-            currency: ISO currency code. Defaults to book's default currency.
-            notes: Optional notes.
-            address: Optional address dict with keys: name, addr1, addr2,
-                     addr3, addr4, phone, fax, email.
-
-        Returns:
-            Dict with guid, id, name, status.
-        """
-        from piecash.business.person import Vendor, Address
-
-        with self.open(readonly=False) as book:
-            if currency:
-                currency_obj = None
-                for c in book.currencies:
-                    if c.mnemonic == currency:
-                        currency_obj = c
-                        break
-                if not currency_obj:
-                    raise ValueError(f"Currency not found: {currency}")
-            else:
-                currency_obj = self._require_default_currency(book)
-
-            addr = None
-            if address:
-                addr = Address(
-                    name=address.get("name", name),
-                    addr1=address.get("addr1", ""),
-                    addr2=address.get("addr2", ""),
-                    addr3=address.get("addr3", ""),
-                    addr4=address.get("addr4", ""),
-                    phone=address.get("phone", ""),
-                    fax=address.get("fax", ""),
-                    email=address.get("email", ""),
-                )
-
-            vendor = Vendor(
-                name=name,
-                currency=currency_obj,
-                notes=notes,
-                address=addr,
-                book=book,
-            )
-            book.save()
-
-            return {
-                "guid": vendor.guid,
-                "id": vendor.id,
-                "name": vendor.name,
-                "status": "created",
-            }
-
-    def list_vendors(
-        self,
-        active_only: bool = True,
-        compact: bool = True,
-    ) -> list[dict] | str:
-        """List all vendors.
-
-        Args:
-            active_only: If True, only return active vendors.
-            compact: If True, return compact one-line-per-vendor string.
-
-        Returns:
-            Compact string or list of dicts.
-        """
-        with self.open() as book:
-            vendors = sorted(book.vendors, key=lambda v: v.name)
-            if active_only:
-                vendors = [v for v in vendors if v.active]
-
-            if compact:
-                lines = [self._vendor_to_compact_line(v) for v in vendors]
-                return "\n".join(lines)
-            else:
-                return [self._vendor_to_dict(v) for v in vendors]
-
-    def get_vendor(self, vendor_id: str) -> dict:
-        """Get vendor details by ID.
-
-        Args:
-            vendor_id: Human-readable vendor ID (e.g., '000001').
-
-        Returns:
-            Dict with full vendor details.
-
-        Raises:
-            ValueError: If vendor not found.
-        """
-        with self.open() as book:
-            vendor = self._find_vendor(book, vendor_id)
-            if not vendor:
-                raise ValueError(f"Vendor not found: {vendor_id}")
-            return self._vendor_to_dict(vendor)
-
-    def create_billterm(
-        self,
-        name: str,
-        due_days: int = 30,
-        description: str = "",
-        discount_days: int = 0,
-        discount_percent: str = "0",
-    ) -> dict:
-        """Create a new billing term.
-
-        Args:
-            name: Billterm name (e.g., 'Net 30').
-            due_days: Number of days until payment is due.
-            description: Optional description.
-            discount_days: Days within which early discount applies.
-            discount_percent: Early payment discount percentage.
-
-        Returns:
-            Dict with guid, name, due_days, status.
-        """
-        import uuid
-        from piecash.business.invoice import Billterm
-
-        discount = Decimal(discount_percent)
-        # Convert discount to num/denom
-        disc_str = str(discount)
-        if "." in disc_str:
-            decimals = len(disc_str.split(".")[1])
-            disc_denom = 10 ** decimals
-            disc_num = int(discount * disc_denom)
-        else:
-            disc_num = int(discount)
-            disc_denom = 1
-
-        with self.open(readonly=False) as book:
-            bt_guid = uuid.uuid4().hex
-
-            book.session.execute(
-                Billterm.__table__.insert().values(
-                    guid=bt_guid,
-                    name=name,
-                    description=description,
-                    refcount=0,
-                    invisible=0,
-                    type="GNC_TERM_TYPE_DAYS",
-                    duedays=due_days,
-                    discountdays=discount_days,
-                    discount_num=disc_num,
-                    discount_denom=disc_denom,
-                    cutoff=0,
-                )
-            )
-            _verify_write(
-                book.session, Billterm.__table__, bt_guid,
-                f"Billterm '{name}'",
-            )
-
-            book.save()
-
-            return {
-                "guid": bt_guid,
-                "name": name,
-                "due_days": due_days,
-                "status": "created",
-            }
-
-    def list_billterms(self, compact: bool = True) -> list[dict] | str:
-        """List all billing terms.
-
-        Args:
-            compact: If True, return compact one-line-per-term string.
-
-        Returns:
-            Compact string or list of dicts.
-        """
-        from piecash.business.invoice import Billterm
-
-        with self.open() as book:
-            terms = book.session.query(Billterm).filter(
-                Billterm.invisible == 0
-            ).order_by(Billterm.name).all()
-
-            if compact:
-                lines = []
-                for t in terms:
-                    lines.append(f"{t.name}\t{t.duedays} days")
-                return "\n".join(lines)
-            else:
-                return [self._billterm_to_dict(t) for t in terms]
-
-    def create_invoice(
-        self,
-        customer_id: str,
-        date_opened: str | None = None,
-        notes: str = "",
-        currency: str | None = None,
-        term: str | None = None,
-        invoice_id: str | None = None,
-    ) -> dict:
-        """Create a customer invoice.
-
-        Args:
-            customer_id: Customer ID (e.g., '000001').
-            date_opened: Date in ISO format. Defaults to today.
-            notes: Optional notes.
-            currency: ISO currency code. Defaults to book's default currency.
-            term: Billterm name (e.g., 'Net 30'). Optional.
-            invoice_id: Custom invoice number. If omitted, auto-generates
-                from the book's invoice counter.
-
-        Returns:
-            Dict with guid, id, customer_id, status.
-        """
-        import uuid
-        from piecash.business.invoice import Invoice, Billterm
-
-        open_date = (
-            datetime.strptime(date_opened, "%Y-%m-%d")
-            if date_opened
-            else datetime.now()
-        )
-
-        with self.open(readonly=False) as book:
-            customer = self._find_customer(book, customer_id)
-            if not customer:
-                raise ValueError(f"Customer not found: {customer_id}")
-
-            if currency:
-                currency_obj = None
-                for c in book.currencies:
-                    if c.mnemonic == currency:
-                        currency_obj = c
-                        break
-                if not currency_obj:
-                    raise ValueError(f"Currency not found: {currency}")
-                currency_guid = currency_obj.guid
-            else:
-                currency_guid = self._require_default_currency(book).guid
-
-            term_guid = None
-            if term:
-                bt = book.session.query(Billterm).filter(
-                    Billterm.name == term, Billterm.invisible == 0
-                ).first()
-                if not bt:
-                    raise ValueError(f"Billterm not found: {term}")
-                term_guid = bt.guid
-
-            if invoice_id is not None:
-                if not invoice_id.strip():
-                    raise ValueError("invoice_id must not be blank")
-                existing = self._find_invoice(book, invoice_id, owner_type=2)
-                if existing:
-                    raise ValueError(
-                        f"Invoice with ID '{invoice_id}' already exists"
-                    )
-            else:
-                cnt = book.counter_invoice + 1
-                book.counter_invoice = cnt
-                invoice_id = f"{cnt:06d}"
-            inv_guid = uuid.uuid4().hex
-
-            book.session.execute(
-                Invoice.__table__.insert().values(
-                    guid=inv_guid,
-                    id=invoice_id,
-                    date_opened=open_date,
-                    date_posted=None,
-                    notes=notes,
-                    active=1,
-                    currency=currency_guid,
-                    owner_type=2,  # Customer
-                    owner_guid=customer.guid,
-                    terms=term_guid,
-                    billing_id="",
-                    post_txn=None,
-                    post_lot=None,
-                    post_acc=None,
-                    billto_type=0,
-                    billto_guid=None,
-                    charge_amt_num=0,
-                    charge_amt_denom=1,
-                )
-            )
-            _verify_write(
-                book.session, Invoice.__table__, inv_guid,
-                f"Invoice '{invoice_id}'",
-            )
-
-            book.save()
-
-            return {
-                "guid": inv_guid,
-                "id": invoice_id,
-                "customer_id": customer_id,
-                "date_opened": str(open_date.date()),
-                "status": "created",
-            }
-
-    def create_bill(
-        self,
-        vendor_id: str,
-        date_opened: str | None = None,
-        notes: str = "",
-        currency: str | None = None,
-        term: str | None = None,
-        bill_id: str | None = None,
-    ) -> dict:
-        """Create a vendor bill.
-
-        Args:
-            vendor_id: Vendor ID (e.g., '000001').
-            date_opened: Date in ISO format. Defaults to today.
-            notes: Optional notes.
-            currency: ISO currency code. Defaults to book's default currency.
-            term: Billterm name (e.g., 'Net 30'). Optional.
-            bill_id: Custom bill number. If omitted, auto-generates
-                from the book's bill counter.
-
-        Returns:
-            Dict with guid, id, vendor_id, status.
-        """
-        import uuid
-        from piecash.business.invoice import Invoice, Billterm
-
-        open_date = (
-            datetime.strptime(date_opened, "%Y-%m-%d")
-            if date_opened
-            else datetime.now()
-        )
-
-        with self.open(readonly=False) as book:
-            vendor = self._find_vendor(book, vendor_id)
-            if not vendor:
-                raise ValueError(f"Vendor not found: {vendor_id}")
-
-            if currency:
-                currency_obj = None
-                for c in book.currencies:
-                    if c.mnemonic == currency:
-                        currency_obj = c
-                        break
-                if not currency_obj:
-                    raise ValueError(f"Currency not found: {currency}")
-                currency_guid = currency_obj.guid
-            else:
-                currency_guid = self._require_default_currency(book).guid
-
-            term_guid = None
-            if term:
-                bt = book.session.query(Billterm).filter(
-                    Billterm.name == term, Billterm.invisible == 0
-                ).first()
-                if not bt:
-                    raise ValueError(f"Billterm not found: {term}")
-                term_guid = bt.guid
-
-            if bill_id is not None:
-                if not bill_id.strip():
-                    raise ValueError("bill_id must not be blank")
-                existing = self._find_invoice(book, bill_id, owner_type=4)
-                if existing:
-                    raise ValueError(
-                        f"Bill with ID '{bill_id}' already exists"
-                    )
-            else:
-                cnt = book.counter_bill + 1
-                book.counter_bill = cnt
-                bill_id = f"{cnt:06d}"
-            inv_guid = uuid.uuid4().hex
-
-            book.session.execute(
-                Invoice.__table__.insert().values(
-                    guid=inv_guid,
-                    id=bill_id,
-                    date_opened=open_date,
-                    date_posted=None,
-                    notes=notes,
-                    active=1,
-                    currency=currency_guid,
-                    owner_type=4,  # Vendor
-                    owner_guid=vendor.guid,
-                    terms=term_guid,
-                    billing_id="",
-                    post_txn=None,
-                    post_lot=None,
-                    post_acc=None,
-                    billto_type=0,
-                    billto_guid=None,
-                    charge_amt_num=0,
-                    charge_amt_denom=1,
-                )
-            )
-            _verify_write(
-                book.session, Invoice.__table__, inv_guid,
-                f"Bill '{bill_id}'",
-            )
-
-            book.save()
-
-            return {
-                "guid": inv_guid,
-                "id": bill_id,
-                "vendor_id": vendor_id,
-                "date_opened": str(open_date.date()),
-                "status": "created",
-            }
-
-    def add_invoice_entry(
-        self,
-        invoice_id: str,
-        account: str,
-        description: str,
-        quantity: str,
-        price: str,
-    ) -> dict:
-        """Add a line item entry to a customer invoice.
-
-        Args:
-            invoice_id: Invoice ID (e.g., '000001').
-            account: Income account full path (e.g., 'Income:Sales').
-            description: Line item description.
-            quantity: Quantity as string (e.g., '1', '2.5').
-            price: Unit price as string (e.g., '100.00').
-
-        Returns:
-            Dict with guid, invoice_id, total, status.
-        """
-        import uuid
-        from piecash.business.invoice import Invoice, Entry
-
-        qty = Decimal(quantity)
-        unit_price = Decimal(price)
-
-        with self.open(readonly=False) as book:
-            inv = self._find_invoice(book, invoice_id, owner_type=2)
-            if not inv:
-                raise ValueError(f"Invoice not found: {invoice_id}")
-            if inv.owner_type != 2:
-                raise ValueError(
-                    f"'{invoice_id}' is a vendor bill, not a customer invoice. "
-                    f"Use add_bill_entry instead."
-                )
-            if inv.date_posted is not None:
-                raise ValueError(
-                    f"Invoice '{invoice_id}' is already posted. "
-                    f"Cannot add entries to posted invoices."
-                )
-
-            acct = self._find_account(book, account)
-            if not acct:
-                raise ValueError(f"Account not found: {account}")
-
-            q_num, q_denom = self._decimal_to_num_denom(qty)
-            p_num, p_denom = self._decimal_to_num_denom(unit_price)
-            entry_guid = uuid.uuid4().hex
-
-            book.session.execute(
-                Entry.__table__.insert().values(
-                    guid=entry_guid,
-                    date=datetime.now(),
-                    date_entered=datetime.now(),
-                    description=description,
-                    action="",
-                    notes="",
-                    quantity_num=q_num,
-                    quantity_denom=q_denom,
-                    i_acct=acct.guid,
-                    i_price_num=p_num,
-                    i_price_denom=p_denom,
-                    i_discount_num=0,
-                    i_discount_denom=1,
-                    invoice=inv.guid,
-                    i_disc_type="",
-                    i_disc_how="",
-                    i_taxable=0,
-                    i_taxincluded=0,
-                    i_taxtable=None,
-                    b_acct=None,
-                    b_price_num=0,
-                    b_price_denom=1,
-                    bill=None,
-                    b_taxable=0,
-                    b_taxincluded=0,
-                    b_taxtable=None,
-                    b_paytype=0,
-                    billable=0,
-                    billto_type=0,
-                    billto_guid=None,
-                    order_guid=None,
-                )
-            )
-            _verify_write(
-                book.session, Entry.__table__, entry_guid,
-                f"Invoice entry '{description}' on invoice {invoice_id}",
-            )
-
-            book.save()
-
-            total = qty * unit_price
-            return {
-                "guid": entry_guid,
-                "invoice_id": invoice_id,
-                "description": description,
-                "quantity": str(qty),
-                "price": str(unit_price),
-                "total": str(total),
-                "status": "created",
-            }
-
-    def add_bill_entry(
-        self,
-        bill_id: str,
-        account: str,
-        description: str,
-        quantity: str,
-        price: str,
-    ) -> dict:
-        """Add a line item entry to a vendor bill.
-
-        Args:
-            bill_id: Bill ID (e.g., '000001').
-            account: Expense account full path (e.g., 'Expenses:Office Supplies').
-            description: Line item description.
-            quantity: Quantity as string (e.g., '1', '2.5').
-            price: Unit price as string (e.g., '50.00').
-
-        Returns:
-            Dict with guid, bill_id, total, status.
-        """
-        import uuid
-        from piecash.business.invoice import Invoice, Entry
-
-        qty = Decimal(quantity)
-        unit_price = Decimal(price)
-
-        with self.open(readonly=False) as book:
-            inv = self._find_invoice(book, bill_id, owner_type=4)
-            if not inv:
-                raise ValueError(f"Bill not found: {bill_id}")
-            if inv.owner_type != 4:
-                raise ValueError(
-                    f"'{bill_id}' is a customer invoice, not a vendor bill. "
-                    f"Use add_invoice_entry instead."
-                )
-            if inv.date_posted is not None:
-                raise ValueError(
-                    f"Bill '{bill_id}' is already posted. "
-                    f"Cannot add entries to posted bills."
-                )
-
-            acct = self._find_account(book, account)
-            if not acct:
-                raise ValueError(f"Account not found: {account}")
-
-            q_num, q_denom = self._decimal_to_num_denom(qty)
-            p_num, p_denom = self._decimal_to_num_denom(unit_price)
-            entry_guid = uuid.uuid4().hex
-
-            book.session.execute(
-                Entry.__table__.insert().values(
-                    guid=entry_guid,
-                    date=datetime.now(),
-                    date_entered=datetime.now(),
-                    description=description,
-                    action="",
-                    notes="",
-                    quantity_num=q_num,
-                    quantity_denom=q_denom,
-                    i_acct=None,
-                    i_price_num=0,
-                    i_price_denom=1,
-                    i_discount_num=0,
-                    i_discount_denom=1,
-                    invoice=None,
-                    i_disc_type="",
-                    i_disc_how="",
-                    i_taxable=0,
-                    i_taxincluded=0,
-                    i_taxtable=None,
-                    b_acct=acct.guid,
-                    b_price_num=p_num,
-                    b_price_denom=p_denom,
-                    bill=inv.guid,
-                    b_taxable=0,
-                    b_taxincluded=0,
-                    b_taxtable=None,
-                    b_paytype=0,
-                    billable=0,
-                    billto_type=0,
-                    billto_guid=None,
-                    order_guid=None,
-                )
-            )
-            _verify_write(
-                book.session, Entry.__table__, entry_guid,
-                f"Bill entry '{description}' on bill {bill_id}",
-            )
-
-            book.save()
-
-            total = qty * unit_price
-            return {
-                "guid": entry_guid,
-                "bill_id": bill_id,
-                "description": description,
-                "quantity": str(qty),
-                "price": str(unit_price),
-                "total": str(total),
-                "status": "created",
-            }
-
-    def list_invoices(
-        self,
-        owner_type: str | None = None,
-        status: str | None = None,
-        compact: bool = True,
-    ) -> list[dict] | str:
-        """List invoices and/or bills.
-
-        Args:
-            owner_type: Filter by type: 'customer', 'vendor', or None for all.
-            status: Filter by status: 'posted', 'open', or None for all.
-            compact: If True, return compact one-line-per-invoice string.
-
-        Returns:
-            Compact string or list of dicts.
-        """
-        from piecash.business.invoice import Invoice
-
-        with self.open() as book:
-            query = book.session.query(Invoice)
-
-            if owner_type == "customer":
-                query = query.filter(Invoice.owner_type == 2)
-            elif owner_type == "vendor":
-                query = query.filter(Invoice.owner_type == 4)
-
-            invoices = query.order_by(Invoice.date_opened.desc()).all()
-
-            if status == "posted":
-                invoices = [i for i in invoices if i.date_posted is not None]
-            elif status == "open":
-                invoices = [i for i in invoices if i.date_posted is None]
-
-            if compact:
-                lines = [self._invoice_to_compact_line(i) for i in invoices]
-                return "\n".join(lines)
-            else:
-                return [self._invoice_to_dict(i) for i in invoices]
-
-    def get_invoice(self, invoice_id: str, owner_type: str | None = None) -> dict:
-        """Get full details for an invoice or bill, including entries.
-
-        Args:
-            invoice_id: Human-readable invoice/bill ID (e.g., '000001').
-            owner_type: Filter by type: 'customer' or 'vendor'.
-                        Useful when invoice and bill share the same ID.
-
-        Returns:
-            Dict with full invoice details and entry list.
-
-        Raises:
-            ValueError: If invoice not found.
-        """
-        from piecash.business.invoice import Invoice, Entry
-        from sqlalchemy import text
-
-        ot = None
-        if owner_type == "customer":
-            ot = 2
-        elif owner_type == "vendor":
-            ot = 4
-
-        with self.open() as book:
-            inv = self._find_invoice(book, invoice_id, owner_type=ot)
-            if not inv:
-                raise ValueError(f"Invoice/bill not found: {invoice_id}")
-
-            is_bill = inv.owner_type == 4
-
-            # Get entries via raw SQL since ORM relationship doesn't
-            # work for vendor bills (bill column is VARCHAR, not FK)
-            if is_bill:
-                rows = book.session.execute(
-                    text("SELECT * FROM entries WHERE bill = :guid"),
-                    {"guid": inv.guid},
-                ).fetchall()
-            else:
-                rows = book.session.execute(
-                    text("SELECT * FROM entries WHERE invoice = :guid"),
-                    {"guid": inv.guid},
-                ).fetchall()
-
-            entries = [self._entry_to_dict(r, is_bill=is_bill) for r in rows]
-
-            # Calculate total
-            total = sum(
-                Decimal(e["quantity"]) * Decimal(e["price"])
-                for e in entries
-            )
-
-            # Look up owner name
-            owner_name = None
-            if is_bill:
-                vendor = self._find_vendor_by_guid(book, inv.owner_guid)
-                if vendor:
-                    owner_name = vendor.name
-            else:
-                customer = self._find_customer_by_guid(book, inv.owner_guid)
-                if customer:
-                    owner_name = customer.name
-
-            result = self._invoice_to_dict(inv, entries=entries)
-            result["total"] = str(total)
-            if owner_name:
-                result["owner_name"] = owner_name
-            return result
-
-    @staticmethod
-    def _find_customer_by_guid(book, guid: str):
-        """Find a customer by GUID."""
-        for c in book.customers:
-            if c.guid == guid:
-                return c
-        return None
-
-    @staticmethod
-    def _find_vendor_by_guid(book, guid: str):
-        """Find a vendor by GUID."""
-        for v in book.vendors:
-            if v.guid == guid:
-                return v
-        return None
-
-    @staticmethod
-    def _write_gncinvoice_slot(book, obj_guid: str, invoice_guid: str):
-        """Write a gncInvoice FRAME+GUID slot linking an object to an invoice.
-
-        GnuCash stores invoice linkage as a two-row structure:
-          Row 1: obj_guid=<parent>, name='gncInvoice', slot_type=9 (FRAME),
-                 guid_val=<frame_guid>
-          Row 2: obj_guid=<frame_guid>, name='invoice', slot_type=5 (GUID),
-                 guid_val=<invoice_guid>
-
-        This is used on both posting transactions and lots to enable
-        GnuCash UI navigation from transaction/lot back to the invoice.
-        """
-        import uuid
-        from piecash.kvp import Slot, KVP_Type
-
-        frame_guid = uuid.uuid4().hex
-        book.session.execute(
-            Slot.__table__.insert().values(
-                obj_guid=obj_guid,
-                name="gncInvoice",
-                slot_type=KVP_Type.KVP_TYPE_FRAME,
-                guid_val=frame_guid,
-            )
-        )
-        _verify_composite_write(
-            book.session, Slot.__table__,
-            {"obj_guid": obj_guid, "name": "gncInvoice"},
-            f"gncInvoice frame slot for {obj_guid[:8]}",
-        )
-        book.session.execute(
-            Slot.__table__.insert().values(
-                obj_guid=frame_guid,
-                name="invoice",
-                slot_type=KVP_Type.KVP_TYPE_GUID,
-                guid_val=invoice_guid,
-            )
-        )
-        _verify_composite_write(
-            book.session, Slot.__table__,
-            {"obj_guid": frame_guid, "name": "invoice"},
-            f"gncInvoice ref slot for frame {frame_guid[:8]}",
-        )
-
-    @staticmethod
-    def _write_gdate_slot(book, obj_guid: str, name: str, date_val: date):
-        """Write a gdate-typed slot on an object.
-
-        Used for trans-date-due and date-posted slots on invoice
-        posting transactions.
-        """
-        from piecash.kvp import Slot, KVP_Type
-
-        book.session.execute(
-            Slot.__table__.insert().values(
-                obj_guid=obj_guid,
-                name=name,
-                slot_type=KVP_Type.KVP_TYPE_GDATE,
-                gdate_val=date_val,
-            )
-        )
-        _verify_composite_write(
-            book.session, Slot.__table__,
-            {"obj_guid": obj_guid, "name": name},
-            f"Gdate slot '{name}' for {obj_guid[:8]}",
-        )
-
-    @staticmethod
-    def _calculate_lot_balance(lot) -> Decimal:
-        """Sum of split values in a lot.
-
-        For A/R lots: positive = outstanding receivable.
-        For A/P lots: negative = outstanding payable.
-        """
-        total = Decimal(0)
-        for split in lot.splits:
-            total += Decimal(str(split.value))
-        return total
-
-    def _get_invoice_entries_and_total(self, book, inv):
-        """Query entries for an invoice/bill and compute total.
-
-        Returns:
-            Tuple of (entries_list, per_account_totals_dict, grand_total).
-            per_account_totals maps account_guid -> Decimal total.
-        """
-        from sqlalchemy import text
-
-        is_bill = inv.owner_type == 4
-        if is_bill:
-            rows = book.session.execute(
-                text("SELECT * FROM entries WHERE bill = :guid"),
-                {"guid": inv.guid},
-            ).fetchall()
-        else:
-            rows = book.session.execute(
-                text("SELECT * FROM entries WHERE invoice = :guid"),
-                {"guid": inv.guid},
-            ).fetchall()
-
-        if not rows:
-            raise ValueError(
-                f"Cannot post: invoice {inv.id} has no entries"
-            )
-
-        # Aggregate totals per account
-        acct_totals: dict[str, Decimal] = {}
-        grand_total = Decimal(0)
-        for row in rows:
-            q_num = row.quantity_num or 0
-            q_denom = row.quantity_denom or 1
-            quantity = Decimal(q_num) / Decimal(q_denom)
-
-            if is_bill:
-                p_num = row.b_price_num or 0
-                p_denom = row.b_price_denom or 1
-                acct_guid = row.b_acct
-            else:
-                p_num = row.i_price_num or 0
-                p_denom = row.i_price_denom or 1
-                acct_guid = row.i_acct
-
-            price = Decimal(p_num) / Decimal(p_denom)
-            entry_total = quantity * price
-            grand_total += entry_total
-            acct_totals[acct_guid] = acct_totals.get(
-                acct_guid, Decimal(0)
-            ) + entry_total
-
-        return rows, acct_totals, grand_total
-
-    def post_invoice(
-        self,
-        invoice_id: str,
-        post_account: str,
-        post_date: str | None = None,
-        due_date: str | None = None,
-        description: str | None = None,
-        owner_type: str | None = None,
-    ) -> dict:
-        """Post a customer invoice or vendor bill.
-
-        Posting creates a transaction in the A/R or A/P account, creates
-        a lot for payment tracking, and marks the invoice as posted.
-
-        Args:
-            invoice_id: Human-readable ID (e.g., '000001').
-            post_account: A/R or A/P account path.
-            post_date: ISO date (YYYY-MM-DD). Defaults to today.
-            due_date: Payment due date (YYYY-MM-DD). Optional.
-            description: Description for the posting transaction.
-            owner_type: 'customer' or 'vendor' for disambiguation.
-
-        Returns:
-            Dict with invoice details, total, transaction_guid, lot_guid.
-
-        Raises:
-            ValueError: If invoice not found, already posted, no entries,
-                        or invalid account type.
-        """
-        from piecash.business.invoice import Invoice
-        from piecash.core.transaction import Lot
-
-        ot = None
-        if owner_type == "customer":
-            ot = 2
-        elif owner_type == "vendor":
-            ot = 4
-
-        parsed_date = (
-            date.fromisoformat(post_date) if post_date
-            else date.today()
-        )
-
-        with self.open(readonly=False) as book:
-            inv = self._find_invoice(book, invoice_id, owner_type=ot)
-            if not inv:
-                raise ValueError(
-                    f"Invoice/bill not found: {invoice_id}"
-                )
-
-            if inv.date_posted is not None:
-                raise ValueError(
-                    f"Invoice {invoice_id} is already posted"
-                )
-
-            is_bill = inv.owner_type == 4
-
-            # Validate post account type
-            post_acct = self._find_account(book, post_account)
-            if not post_acct:
-                raise ValueError(
-                    f"Account not found: {post_account}"
-                )
-            expected_type = "PAYABLE" if is_bill else "RECEIVABLE"
-            if post_acct.type != expected_type:
-                raise ValueError(
-                    f"Post account must be {expected_type}, "
-                    f"got {post_acct.type}"
-                )
-
-            # Get entries and totals
-            _, acct_totals, grand_total = (
-                self._get_invoice_entries_and_total(book, inv)
-            )
-
-            # Create lot for payment tracking
-            lot = Lot(
-                title=f"Invoice {inv.id}",
-                account=post_acct,
-                is_closed=0,
-            )
-            book.session.add(lot)
-
-            # Resolve owner name for transaction description
-            # GnuCash UI uses the customer/vendor name, not "Invoice NNNNNN"
-            if is_bill:
-                owner = self._find_vendor_by_guid(book, inv.owner_guid)
-            else:
-                owner = self._find_customer_by_guid(
-                    book, inv.owner_guid
-                )
-            owner_name = owner.name if owner else f"Invoice {inv.id}"
-            txn_desc = description or owner_name
-
-            # Parse due date if provided
-            parsed_due = (
-                date.fromisoformat(due_date) if due_date else None
-            )
-
-            # Build transaction splits
-            # For customer invoice: A/R debit (positive), income credit (negative)
-            # For vendor bill: A/P credit (negative), expense debit (positive)
-            piecash_splits = []
-
-            if is_bill:
-                # A/P split: negative (credit)
-                ar_ap_split = piecash.Split(
-                    account=post_acct,
-                    value=-grand_total,
-                    quantity=-grand_total,
-                    memo="",
-                    action="Invoice",
-                    reconcile_date=datetime(1970, 1, 1),
-                )
-            else:
-                # A/R split: positive (debit)
-                ar_ap_split = piecash.Split(
-                    account=post_acct,
-                    value=grand_total,
-                    quantity=grand_total,
-                    memo="",
-                    action="Invoice",
-                    reconcile_date=datetime(1970, 1, 1),
-                )
-            piecash_splits.append(ar_ap_split)
-
-            # One split per distinct entry account
-            for acct_guid, acct_total in acct_totals.items():
-                entry_acct = None
-                for a in book.accounts:
-                    if a.guid == acct_guid:
-                        entry_acct = a
-                        break
-                if not entry_acct:
-                    raise ValueError(
-                        f"Entry account not found: {acct_guid}"
-                    )
-
-                if is_bill:
-                    # Expense: positive (debit)
-                    split_value = acct_total
-                else:
-                    # Income: negative (credit)
-                    split_value = -acct_total
-
-                piecash_splits.append(
-                    piecash.Split(
-                        account=entry_acct,
-                        value=split_value,
-                        quantity=split_value,
-                        memo="",
-                    )
-                )
-
-            # Create posting transaction
-            # num = invoice ID, matching GnuCash UI behavior
-            txn = piecash.Transaction(
-                currency=inv.currency,
-                description=txn_desc,
-                post_date=parsed_date,
-                num=inv.id,
-                splits=piecash_splits,
-            )
-
-            # Assign A/R or A/P split to the lot
-            ar_ap_split.lot = lot
-
-            # Update invoice record with posting info via ORM
-            inv.date_posted = datetime.combine(
-                parsed_date, datetime.min.time()
-            )
-            inv.post_txn = txn
-            inv.post_lot = lot
-            inv.post_account = post_acct
-
-            # Flush to assign GUIDs before writing slots
-            book.flush()
-
-            # Write metadata slots matching GnuCash UI behavior.
-            # These enable GnuCash to identify invoice-generated
-            # transactions and navigate between invoices/lots/txns.
-            txn["trans-txn-type"] = "I"
-            txn["trans-read-only"] = (
-                "Generated from an invoice. "
-                "Try unposting the invoice."
-            )
-            self._write_gncinvoice_slot(
-                book, txn.guid, inv.guid
-            )
-            self._write_gncinvoice_slot(
-                book, lot.guid, inv.guid
-            )
-            if parsed_due:
-                self._write_gdate_slot(
-                    book, txn.guid, "trans-date-due", parsed_due
-                )
-
-            book.save()
-
-            # Capture return values before session closes
-            result = {
-                "id": inv.id,
-                "type": "bill" if is_bill else "invoice",
-                "status": "posted",
-                "total": str(grand_total),
-                "post_date": str(parsed_date),
-                "transaction_guid": txn.guid,
-                "lot_guid": lot.guid,
-                "post_account": post_acct.fullname,
-            }
-
-        return result
-
-    def pay_invoice(
-        self,
-        invoice_id: str,
-        payment_account: str,
-        amount: str,
-        payment_date: str | None = None,
-        description: str | None = None,
-        owner_type: str | None = None,
-    ) -> dict:
-        """Record a payment against a posted invoice or bill.
-
-        Creates a payment transaction and assigns the A/R or A/P split
-        to the invoice's lot for balance tracking. Partial payments
-        are supported.
-
-        Args:
-            invoice_id: Human-readable ID (e.g., '000001').
-            payment_account: Bank or cash account path.
-            amount: Payment amount as decimal string (e.g., '500.00').
-            payment_date: ISO date (YYYY-MM-DD). Defaults to today.
-            description: Description for the payment transaction.
-            owner_type: 'customer' or 'vendor' for disambiguation.
-
-        Returns:
-            Dict with payment details and remaining balance.
-
-        Raises:
-            ValueError: If invoice not found, not posted, or invalid account.
-        """
-        ot = None
-        if owner_type == "customer":
-            ot = 2
-        elif owner_type == "vendor":
-            ot = 4
-
-        payment_amount = Decimal(amount)
-        if payment_amount <= 0:
-            raise ValueError("Payment amount must be positive")
-
-        parsed_date = (
-            date.fromisoformat(payment_date) if payment_date
-            else date.today()
-        )
-
-        with self.open(readonly=False) as book:
-            inv = self._find_invoice(book, invoice_id, owner_type=ot)
-            if not inv:
-                raise ValueError(
-                    f"Invoice/bill not found: {invoice_id}"
-                )
-
-            if inv.date_posted is None:
-                raise ValueError(
-                    f"Invoice {invoice_id} is not posted — "
-                    f"post it before recording payment"
-                )
-
-            is_bill = inv.owner_type == 4
-
-            # Find payment account
-            pay_acct = self._find_account(book, payment_account)
-            if not pay_acct:
-                raise ValueError(
-                    f"Account not found: {payment_account}"
-                )
-
-            # Find the post account (A/R or A/P) from invoice
-            post_acct = None
-            post_acc_guid = inv.post_acc_guid
-            for a in book.accounts:
-                if a.guid == post_acc_guid:
-                    post_acct = a
-                    break
-            if not post_acct:
-                raise ValueError(
-                    f"Post account not found for invoice {invoice_id}"
-                )
-
-            # Find the lot
-            lot_guid = inv.post_lot_guid
-            lot_obj = None
-            for lot in post_acct.lots:
-                if lot.guid == lot_guid:
-                    lot_obj = lot
-                    break
-            if not lot_obj:
-                raise ValueError(
-                    f"Lot not found for invoice {invoice_id}"
-                )
-
-            # Resolve owner name for transaction description
-            # GnuCash UI uses the customer/vendor name
-            if is_bill:
-                owner = self._find_vendor_by_guid(
-                    book, inv.owner_guid
-                )
-            else:
-                owner = self._find_customer_by_guid(
-                    book, inv.owner_guid
-                )
-            owner_name = owner.name if owner else ""
-            txn_desc = description or owner_name
-
-            # Build payment splits
-            if is_bill:
-                # Pay vendor bill: debit A/P (positive), credit bank (negative)
-                ar_ap_split = piecash.Split(
-                    account=post_acct,
-                    value=payment_amount,
-                    quantity=payment_amount,
-                    memo="",
-                    action="Payment",
-                )
-                bank_split = piecash.Split(
-                    account=pay_acct,
-                    value=-payment_amount,
-                    quantity=-payment_amount,
-                    memo="",
-                )
-            else:
-                # Receive customer payment: credit A/R (negative), debit bank (positive)
-                ar_ap_split = piecash.Split(
-                    account=post_acct,
-                    value=-payment_amount,
-                    quantity=-payment_amount,
-                    memo="",
-                    action="Payment",
-                )
-                bank_split = piecash.Split(
-                    account=pay_acct,
-                    value=payment_amount,
-                    quantity=payment_amount,
-                    memo="",
-                )
-
-            # Create payment transaction
-            txn = piecash.Transaction(
-                currency=inv.currency,
-                description=txn_desc,
-                post_date=parsed_date,
-                num="",
-                splits=[ar_ap_split, bank_split],
-            )
-
-            # Assign A/R or A/P split to the invoice's lot
-            ar_ap_split.lot = lot_obj
-
-            # Flush to assign GUIDs before writing slots
-            book.flush()
-
-            # Mark as payment transaction for GnuCash UI
-            txn["trans-txn-type"] = "P"
-
-            # Auto-close lot if fully paid
-            remaining = self._calculate_lot_balance(lot_obj)
-            if remaining == Decimal(0):
-                lot_obj.is_closed = -1
-
-            book.save()
-
-            result = {
-                "id": inv.id,
-                "type": "bill" if is_bill else "invoice",
-                "status": "paid",
-                "amount_paid": str(payment_amount),
-                "remaining_balance": str(abs(remaining)),
-                "transaction_guid": txn.guid,
-                "payment_account": pay_acct.fullname,
-                "payment_date": str(parsed_date),
-            }
-
-        return result
-
-    def delete_invoice(self, invoice_id: str) -> dict:
-        """Delete an unposted customer invoice.
-
-        Automatically removes associated entries. Posted invoices cannot
-        be deleted — void them or issue a credit note instead.
-
-        Args:
-            invoice_id: Invoice ID (e.g., '000001' or 'INV-2026-001').
-
-        Returns:
-            Dict with id, guid, entries_deleted, status.
-
-        Raises:
-            ValueError: If invoice not found or is posted.
-        """
-        return self._delete_invoice_or_bill(invoice_id, owner_type=2)
-
-    def delete_bill(self, bill_id: str) -> dict:
-        """Delete an unposted vendor bill.
-
-        Automatically removes associated entries. Posted bills cannot
-        be deleted — void them or issue a credit note instead.
-
-        Args:
-            bill_id: Bill ID (e.g., '000001' or 'BILL-2026-001').
-
-        Returns:
-            Dict with id, guid, entries_deleted, status.
-
-        Raises:
-            ValueError: If bill not found or is posted.
-        """
-        return self._delete_invoice_or_bill(bill_id, owner_type=4)
-
-    def _delete_invoice_or_bill(self, doc_id: str, owner_type: int) -> dict:
-        """Shared implementation for delete_invoice and delete_bill."""
-        from sqlalchemy import text
-        from piecash.business.invoice import Invoice
-
-        type_label = "Bill" if owner_type == 4 else "Invoice"
-        entry_fk = "bill" if owner_type == 4 else "invoice"
-
-        with self.open(readonly=False) as book:
-            inv = self._find_invoice(book, doc_id, owner_type=owner_type)
-            if not inv:
-                raise ValueError(f"{type_label} not found: {doc_id}")
-
-            if inv.date_posted is not None:
-                raise ValueError(
-                    f"Cannot delete posted {type_label.lower()} '{doc_id}'. "
-                    f"Void it or issue a credit note instead."
-                )
-
-            inv_guid = inv.guid
-
-            # Delete entries
-            entry_count = book.session.execute(
-                text(f"SELECT count(*) FROM entries WHERE {entry_fk} = :guid"),
-                {"guid": inv_guid},
-            ).scalar()
-            if entry_count:
-                book.session.execute(
-                    text(f"DELETE FROM entries WHERE {entry_fk} = :guid"),
-                    {"guid": inv_guid},
-                )
-
-            # Delete the invoice row
-            book.session.execute(
-                text("DELETE FROM invoices WHERE guid = :guid"),
-                {"guid": inv_guid},
-            )
-
-            book.save()
-
-            return {
-                "id": doc_id,
-                "guid": inv_guid,
-                "type": type_label.lower(),
-                "entries_deleted": entry_count,
-                "status": "deleted",
-            }
-
-    def delete_customer(self, customer_id: str) -> dict:
-        """Delete a customer with no invoices.
-
-        Customers with any invoices (posted or unposted) cannot be deleted.
-        Delete the invoices first, or void posted ones.
-
-        Args:
-            customer_id: Customer ID (e.g., '000001').
-
-        Returns:
-            Dict with id, guid, name, status.
-
-        Raises:
-            ValueError: If customer not found or has invoices.
-        """
-        return self._delete_customer_or_vendor(customer_id, is_vendor=False)
-
-    def delete_vendor(self, vendor_id: str) -> dict:
-        """Delete a vendor with no bills.
-
-        Vendors with any bills (posted or unposted) cannot be deleted.
-        Delete the bills first, or void posted ones.
-
-        Args:
-            vendor_id: Vendor ID (e.g., '000001').
-
-        Returns:
-            Dict with id, guid, name, status.
-
-        Raises:
-            ValueError: If vendor not found or has bills.
-        """
-        return self._delete_customer_or_vendor(vendor_id, is_vendor=True)
-
-    def _delete_customer_or_vendor(self, entity_id: str, is_vendor: bool) -> dict:
-        """Shared implementation for delete_customer and delete_vendor."""
-        from sqlalchemy import text
-        from piecash.business.invoice import Invoice
-
-        if is_vendor:
-            type_label = "Vendor"
-            owner_type = 4
-            doc_label = "bills"
-        else:
-            type_label = "Customer"
-            owner_type = 2
-            doc_label = "invoices"
-
-        with self.open(readonly=False) as book:
-            if is_vendor:
-                entity = self._find_vendor(book, entity_id)
-            else:
-                entity = self._find_customer(book, entity_id)
-
-            if not entity:
-                raise ValueError(f"{type_label} not found: {entity_id}")
-
-            entity_guid = entity.guid
-            entity_name = entity.name
-
-            # Check for any invoices/bills
-            invoices = book.session.query(Invoice).filter(
-                Invoice.owner_guid == entity_guid,
-                Invoice.owner_type == owner_type,
-            ).all()
-
-            if invoices:
-                posted = [i for i in invoices if i.date_posted is not None]
-                unposted = [i for i in invoices if i.date_posted is None]
-
-                if posted:
-                    posted_ids = ", ".join(i.id for i in posted)
-                    raise ValueError(
-                        f"Cannot delete {type_label.lower()} with posted "
-                        f"{doc_label}: {posted_ids}. "
-                        f"Void them or issue credit notes first."
-                    )
-                else:
-                    unposted_ids = ", ".join(i.id for i in unposted)
-                    raise ValueError(
-                        f"Cannot delete {type_label.lower()} with "
-                        f"{doc_label}: {unposted_ids}. "
-                        f"Delete the {doc_label} first."
-                    )
-
-            # Delete slots for this entity
-            book.session.execute(
-                text("DELETE FROM slots WHERE obj_guid = :guid"),
-                {"guid": entity_guid},
-            )
-
-            # Delete the entity
-            book.session.delete(entity)
-            book.save()
-
-            return {
-                "id": entity_id,
-                "guid": entity_guid,
-                "name": entity_name,
-                "type": type_label.lower(),
-                "status": "deleted",
-            }
-
-    def get_outstanding_invoices(
-        self,
-        owner_type: str | None = None,
-        customer_id: str | None = None,
-        vendor_id: str | None = None,
-    ) -> list[dict]:
-        """Get all posted invoices/bills with outstanding balances.
-
-        Args:
-            owner_type: Filter by 'customer' or 'vendor'. Omit for all.
-            customer_id: Filter by specific customer ID.
-            vendor_id: Filter by specific vendor ID.
-
-        Returns:
-            List of dicts with invoice details and balance info.
-        """
-        from piecash.business.invoice import Invoice
-
-        with self.open() as book:
-            query = book.session.query(Invoice).filter(
-                Invoice.date_posted.isnot(None)
-            )
-
-            if owner_type == "customer":
-                query = query.filter(Invoice.owner_type == 2)
-            elif owner_type == "vendor":
-                query = query.filter(Invoice.owner_type == 4)
-
-            # Filter by specific customer or vendor
-            if customer_id:
-                customer = None
-                for c in book.customers:
-                    if c.id == customer_id:
-                        customer = c
-                        break
-                if not customer:
-                    raise ValueError(
-                        f"Customer not found: {customer_id}"
-                    )
-                query = query.filter(
-                    Invoice.owner_guid == customer.guid
-                )
-
-            if vendor_id:
-                vendor = None
-                for v in book.vendors:
-                    if v.id == vendor_id:
-                        vendor = v
-                        break
-                if not vendor:
-                    raise ValueError(
-                        f"Vendor not found: {vendor_id}"
-                    )
-                query = query.filter(
-                    Invoice.owner_guid == vendor.guid
-                )
-
-            invoices = query.order_by(
-                Invoice.date_posted.desc()
-            ).all()
-
-            results = []
-            for inv in invoices:
-                is_bill = inv.owner_type == 4
-
-                # Find lot and calculate balance
-                post_acc_guid = inv.post_acc_guid
-                post_acct = None
-                for a in book.accounts:
-                    if a.guid == post_acc_guid:
-                        post_acct = a
-                        break
-                if not post_acct:
-                    continue
-
-                lot_obj = None
-                for lot in post_acct.lots:
-                    if lot.guid == inv.post_lot_guid:
-                        lot_obj = lot
-                        break
-                if not lot_obj:
-                    continue
-
-                balance = self._calculate_lot_balance(lot_obj)
-                # Skip fully paid (balance is zero)
-                if balance == Decimal(0):
-                    continue
-
-                # Calculate original total from entries
-                try:
-                    _, _, grand_total = (
-                        self._get_invoice_entries_and_total(book, inv)
-                    )
-                except ValueError:
-                    grand_total = abs(balance)
-
-                amount_paid = grand_total - abs(balance)
-
-                # Owner name
-                owner_name = None
-                if is_bill:
-                    v = self._find_vendor_by_guid(
-                        book, inv.owner_guid
-                    )
-                    if v:
-                        owner_name = v.name
-                else:
-                    c = self._find_customer_by_guid(
-                        book, inv.owner_guid
-                    )
-                    if c:
-                        owner_name = c.name
-
-                results.append({
-                    "id": inv.id,
-                    "type": "bill" if is_bill else "invoice",
-                    "owner_name": owner_name,
-                    "date_posted": (
-                        str(inv.date_posted.date())
-                        if inv.date_posted else None
-                    ),
-                    "original_amount": str(grand_total),
-                    "amount_paid": str(amount_paid),
-                    "amount_due": str(abs(balance)),
-                })
-
-        return results
-
-    def vendor_spending_report(
-        self,
-        start_date: str,
-        end_date: str,
-        vendor_id: str | None = None,
-    ) -> dict:
-        """Get spending breakdown by vendor for a period.
-
-        Analyzes posted vendor bills to show total billed, paid,
-        and outstanding amounts per vendor.
-
-        Args:
-            start_date: Start of period (YYYY-MM-DD).
-            end_date: End of period (YYYY-MM-DD).
-            vendor_id: Optional filter to specific vendor.
-
-        Returns:
-            Dict with per-vendor breakdown and grand totals.
-        """
-        from piecash.business.invoice import Invoice
-
-        parsed_start = date.fromisoformat(start_date)
-        parsed_end = date.fromisoformat(end_date)
-
-        with self.open() as book:
-            query = book.session.query(Invoice).filter(
-                Invoice.owner_type == 4,
-                Invoice.date_posted.isnot(None),
-            )
-
-            if vendor_id:
-                vendor = None
-                for v in book.vendors:
-                    if v.id == vendor_id:
-                        vendor = v
-                        break
-                if not vendor:
-                    raise ValueError(
-                        f"Vendor not found: {vendor_id}"
-                    )
-                query = query.filter(
-                    Invoice.owner_guid == vendor.guid
-                )
-
-            bills = query.all()
-
-            # Filter by date range (date_posted is datetime)
-            bills = [
-                b for b in bills
-                if parsed_start <= b.date_posted.date() <= parsed_end
-            ]
-
-            # Aggregate by vendor
-            vendor_data: dict[str, dict] = {}
-            for bill in bills:
-                # Get vendor name
-                v = self._find_vendor_by_guid(
-                    book, bill.owner_guid
-                )
-                v_name = v.name if v else "Unknown"
-                v_id = v.id if v else ""
-
-                if v_name not in vendor_data:
-                    vendor_data[v_name] = {
-                        "vendor_id": v_id,
-                        "vendor_name": v_name,
-                        "total_billed": Decimal(0),
-                        "total_paid": Decimal(0),
-                        "outstanding": Decimal(0),
-                        "bill_count": 0,
-                    }
-
-                # Calculate bill total from entries
-                try:
-                    _, _, total = (
-                        self._get_invoice_entries_and_total(
-                            book, bill
-                        )
-                    )
-                except ValueError:
-                    total = Decimal(0)
-
-                # Calculate outstanding from lot
-                balance = Decimal(0)
-                post_acct = None
-                for a in book.accounts:
-                    if a.guid == bill.post_acc_guid:
-                        post_acct = a
-                        break
-                if post_acct:
-                    for lot in post_acct.lots:
-                        if lot.guid == bill.post_lot_guid:
-                            balance = self._calculate_lot_balance(
-                                lot
-                            )
-                            break
-
-                outstanding = abs(balance)
-                paid = total - outstanding
-
-                vendor_data[v_name]["total_billed"] += total
-                vendor_data[v_name]["total_paid"] += paid
-                vendor_data[v_name]["outstanding"] += outstanding
-                vendor_data[v_name]["bill_count"] += 1
-
-            # Convert Decimals to strings
-            vendors_list = []
-            grand_billed = Decimal(0)
-            grand_paid = Decimal(0)
-            grand_outstanding = Decimal(0)
-
-            for vd in vendor_data.values():
-                grand_billed += vd["total_billed"]
-                grand_paid += vd["total_paid"]
-                grand_outstanding += vd["outstanding"]
-                vendors_list.append({
-                    "vendor_id": vd["vendor_id"],
-                    "vendor_name": vd["vendor_name"],
-                    "total_billed": str(vd["total_billed"]),
-                    "total_paid": str(vd["total_paid"]),
-                    "outstanding": str(vd["outstanding"]),
-                    "bill_count": vd["bill_count"],
-                })
-
-            # Sort by total billed descending
-            vendors_list.sort(
-                key=lambda x: Decimal(x["total_billed"]),
-                reverse=True,
-            )
-
-        return {
-            "period": {
-                "start": start_date,
-                "end": end_date,
-            },
-            "vendors": vendors_list,
-            "totals": {
-                "total_billed": str(grand_billed),
-                "total_paid": str(grand_paid),
-                "outstanding": str(grand_outstanding),
-                "vendor_count": len(vendors_list),
-                "bill_count": sum(
-                    v["bill_count"] for v in vendors_list
-                ),
-            },
-        }
+    # Business methods moved to book/business.py as BusinessMixin.
