+++ src/gnucash_mcp/tools/business.py
@@ -1,0 +1,536 @@
+"""Business tools: customers, vendors, billterms, invoices, bills.
+
+Registered only when the 'business' module is enabled via --modules.
+"""
+
+import json
+
+from gnucash_mcp.logging_config import audit_log
+from gnucash_mcp.tools._helpers import _json, safe_tool
+
+
+def register(mcp, get_book) -> None:
+    """Attach business tools to the FastMCP server."""
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="create", entity_type="customer")
+    def create_customer(
+        name: str,
+        currency: str | None = None,
+        notes: str = "",
+        address: dict | None = None,
+    ) -> str:
+        """Create a new customer.
+
+        Args:
+            name: Customer name (e.g., "Acme Corp").
+            currency: ISO currency code (e.g., "USD", "EUR").
+                      Defaults to book's default currency.
+            notes: Optional notes.
+            address: Optional address with keys: name, addr1, addr2,
+                     addr3, addr4, phone, fax, email.
+        """
+        book = get_book()
+        result = book.create_customer(
+            name=name, currency=currency, notes=notes, address=address,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="read")
+    def list_customers(
+        active_only: bool = True,
+        verbose: bool = False,
+    ) -> str:
+        """List all customers.
+
+        Returns a compact one-line-per-customer format by default.
+        Use verbose=true for full JSON with guid, address, notes, etc.
+
+        Args:
+            active_only: If True, only show active customers. Default True.
+            verbose: If true, return full JSON details for each customer.
+        """
+        book = get_book()
+        result = book.list_customers(active_only=active_only, compact=not verbose)
+        if verbose:
+            return json.dumps(result, indent=2)
+        return result
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="read")
+    def get_customer(
+        id: str,
+    ) -> str:
+        """Get details for a specific customer by ID.
+
+        Args:
+            id: Customer ID (e.g., "000001"). This is the human-readable
+                ID shown in GnuCash, not the internal GUID.
+        """
+        book = get_book()
+        result = book.get_customer(customer_id=id)
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="create", entity_type="vendor")
+    def create_vendor(
+        name: str,
+        currency: str | None = None,
+        notes: str = "",
+        address: dict | None = None,
+    ) -> str:
+        """Create a new vendor.
+
+        Args:
+            name: Vendor name (e.g., "Office Depot").
+            currency: ISO currency code (e.g., "USD", "EUR").
+                      Defaults to book's default currency.
+            notes: Optional notes.
+            address: Optional address with keys: name, addr1, addr2,
+                     addr3, addr4, phone, fax, email.
+        """
+        book = get_book()
+        result = book.create_vendor(
+            name=name, currency=currency, notes=notes, address=address,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="read")
+    def list_vendors(
+        active_only: bool = True,
+        verbose: bool = False,
+    ) -> str:
+        """List all vendors.
+
+        Returns a compact one-line-per-vendor format by default.
+        Use verbose=true for full JSON with guid, address, notes, etc.
+
+        Args:
+            active_only: If True, only show active vendors. Default True.
+            verbose: If true, return full JSON details for each vendor.
+        """
+        book = get_book()
+        result = book.list_vendors(active_only=active_only, compact=not verbose)
+        if verbose:
+            return json.dumps(result, indent=2)
+        return result
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="read")
+    def get_vendor(
+        id: str,
+    ) -> str:
+        """Get details for a specific vendor by ID.
+
+        Args:
+            id: Vendor ID (e.g., "000001"). This is the human-readable
+                ID shown in GnuCash, not the internal GUID.
+        """
+        book = get_book()
+        result = book.get_vendor(vendor_id=id)
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="create", entity_type="billterm")
+    def create_billterm(
+        name: str,
+        due_days: int = 30,
+        description: str = "",
+        discount_days: int = 0,
+        discount_percent: str = "0",
+    ) -> str:
+        """Create a new billing term.
+
+        Args:
+            name: Billterm name (e.g., "Net 30").
+            due_days: Number of days until payment is due. Default 30.
+            description: Optional description.
+            discount_days: Days within which early discount applies.
+            discount_percent: Early payment discount percentage (e.g., "2" for 2%).
+        """
+        book = get_book()
+        result = book.create_billterm(
+            name=name, due_days=due_days, description=description,
+            discount_days=discount_days, discount_percent=discount_percent,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="read")
+    def list_billterms(
+        verbose: bool = False,
+    ) -> str:
+        """List all billing terms.
+
+        Returns a compact one-line-per-term format by default.
+        Use verbose=true for full JSON with guid, discount details, etc.
+
+        Args:
+            verbose: If true, return full JSON details for each billing term.
+        """
+        book = get_book()
+        result = book.list_billterms(compact=not verbose)
+        if verbose:
+            return json.dumps(result, indent=2)
+        return result
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="create", entity_type="invoice")
+    def create_invoice(
+        customer_id: str,
+        date_opened: str | None = None,
+        notes: str = "",
+        currency: str | None = None,
+        term: str | None = None,
+        invoice_id: str | None = None,
+    ) -> str:
+        """Create a customer invoice.
+
+        Args:
+            customer_id: Customer ID (e.g., "000001").
+            date_opened: Date in ISO format (YYYY-MM-DD). Defaults to today.
+            notes: Optional notes.
+            currency: ISO currency code. Defaults to book's default currency.
+            term: Billterm name (e.g., "Net 30"). Optional.
+            invoice_id: Custom invoice number (e.g., "INV-2026-001"). If omitted,
+                auto-generates from the book's invoice counter.
+        """
+        book = get_book()
+        result = book.create_invoice(
+            customer_id=customer_id, date_opened=date_opened,
+            notes=notes, currency=currency, term=term,
+            invoice_id=invoice_id,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="create", entity_type="bill")
+    def create_bill(
+        vendor_id: str,
+        date_opened: str | None = None,
+        notes: str = "",
+        currency: str | None = None,
+        term: str | None = None,
+        bill_id: str | None = None,
+    ) -> str:
+        """Create a vendor bill.
+
+        Args:
+            vendor_id: Vendor ID (e.g., "000001").
+            date_opened: Date in ISO format (YYYY-MM-DD). Defaults to today.
+            notes: Optional notes.
+            currency: ISO currency code. Defaults to book's default currency.
+            term: Billterm name (e.g., "Net 30"). Optional.
+            bill_id: Custom bill number (e.g., "BILL-2026-001"). If omitted,
+                auto-generates from the book's bill counter.
+        """
+        book = get_book()
+        result = book.create_bill(
+            vendor_id=vendor_id, date_opened=date_opened,
+            notes=notes, currency=currency, term=term,
+            bill_id=bill_id,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="create", entity_type="entry")
+    def add_invoice_entry(
+        invoice_id: str,
+        account: str,
+        description: str,
+        quantity: str,
+        price: str,
+    ) -> str:
+        """Add a line item to a customer invoice.
+
+        The invoice must not be posted yet. Entries represent individual
+        goods or services being billed.
+
+        Args:
+            invoice_id: Invoice ID (e.g., "000001").
+            account: Income account path (e.g., "Income:Sales").
+            description: Line item description.
+            quantity: Quantity as decimal string (e.g., "1", "2.5").
+            price: Unit price as decimal string (e.g., "100.00").
+        """
+        book = get_book()
+        result = book.add_invoice_entry(
+            invoice_id=invoice_id, account=account,
+            description=description, quantity=quantity, price=price,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="create", entity_type="entry")
+    def add_bill_entry(
+        bill_id: str,
+        account: str,
+        description: str,
+        quantity: str,
+        price: str,
+    ) -> str:
+        """Add a line item to a vendor bill.
+
+        The bill must not be posted yet. Entries represent individual
+        goods or services being billed.
+
+        Args:
+            bill_id: Bill ID (e.g., "000001").
+            account: Expense account path (e.g., "Expenses:Office Supplies").
+            description: Line item description.
+            quantity: Quantity as decimal string (e.g., "1", "2.5").
+            price: Unit price as decimal string (e.g., "50.00").
+        """
+        book = get_book()
+        result = book.add_bill_entry(
+            bill_id=bill_id, account=account,
+            description=description, quantity=quantity, price=price,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="read")
+    def list_invoices(
+        owner_type: str | None = None,
+        status: str | None = None,
+        verbose: bool = False,
+    ) -> str:
+        """List invoices and/or vendor bills.
+
+        Returns a compact one-line-per-invoice format by default.
+        Use verbose=true for full JSON with GUIDs, dates, notes, etc.
+
+        Args:
+            owner_type: Filter by type: "customer" for invoices,
+                        "vendor" for bills, or omit for all.
+            status: Filter by status: "posted" or "open", or omit for all.
+            verbose: If true, return full JSON details.
+        """
+        book = get_book()
+        result = book.list_invoices(
+            owner_type=owner_type, status=status, compact=not verbose,
+        )
+        if verbose:
+            return json.dumps(result, indent=2)
+        return result
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="read")
+    def get_invoice(
+        id: str,
+        owner_type: str | None = None,
+    ) -> str:
+        """Get full details for an invoice or bill, including line items.
+
+        Works for both customer invoices and vendor bills. Returns all
+        entries with quantities, prices, and totals.
+
+        Args:
+            id: Invoice or bill ID (e.g., "000001"). This is the
+                human-readable ID, not the internal GUID.
+            owner_type: Filter by type: "customer" for invoices,
+                        "vendor" for bills. Useful when an invoice and
+                        bill share the same ID (independent counters).
+        """
+        book = get_book()
+        result = book.get_invoice(invoice_id=id, owner_type=owner_type)
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="post", entity_type="invoice")
+    def post_invoice(
+        id: str,
+        post_account: str,
+        post_date: str | None = None,
+        due_date: str | None = None,
+        description: str | None = None,
+        owner_type: str | None = None,
+    ) -> str:
+        """Post a customer invoice or vendor bill.
+
+        Posting creates a transaction in the A/R or A/P account and makes
+        the invoice official. Once posted, entries cannot be added.
+
+        Args:
+            id: Invoice or bill ID (e.g., "000001").
+            post_account: A/R or A/P account path (e.g., "Assets:Accounts Receivable").
+            post_date: Date in ISO format (YYYY-MM-DD). Defaults to today.
+            due_date: Payment due date (YYYY-MM-DD). Optional.
+            description: Description for the posting transaction. Optional.
+            owner_type: "customer" or "vendor" for disambiguation when IDs collide.
+        """
+        book = get_book()
+        result = book.post_invoice(
+            invoice_id=id,
+            post_account=post_account,
+            post_date=post_date,
+            due_date=due_date,
+            description=description,
+            owner_type=owner_type,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="pay", entity_type="invoice")
+    def pay_invoice(
+        id: str,
+        payment_account: str,
+        amount: str,
+        payment_date: str | None = None,
+        description: str | None = None,
+        owner_type: str | None = None,
+    ) -> str:
+        """Record a payment against a posted invoice or bill.
+
+        Creates a payment transaction from the specified bank/cash account
+        to the invoice's A/R or A/P account. Partial payments are supported.
+
+        Args:
+            id: Invoice or bill ID (e.g., "000001").
+            payment_account: Bank or cash account for payment (e.g., "Assets:Checking").
+            amount: Payment amount as decimal string (e.g., "500.00").
+            payment_date: Payment date (YYYY-MM-DD). Defaults to today.
+            description: Description for the payment transaction. Optional.
+            owner_type: "customer" or "vendor" for disambiguation.
+        """
+        book = get_book()
+        result = book.pay_invoice(
+            invoice_id=id,
+            payment_account=payment_account,
+            amount=amount,
+            payment_date=payment_date,
+            description=description,
+            owner_type=owner_type,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="delete", entity_type="invoice")
+    def delete_invoice(invoice_id: str) -> str:
+        """Delete an unposted customer invoice.
+
+        Automatically removes associated entries (line items). Posted invoices
+        cannot be deleted — void them or issue a credit note instead.
+
+        Args:
+            invoice_id: Invoice ID (e.g., "000001" or "INV-2026-001").
+        """
+        book = get_book()
+        result = book.delete_invoice(invoice_id=invoice_id)
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="delete", entity_type="bill")
+    def delete_bill(bill_id: str) -> str:
+        """Delete an unposted vendor bill.
+
+        Automatically removes associated entries (line items). Posted bills
+        cannot be deleted — void them or issue a credit note instead.
+
+        Args:
+            bill_id: Bill ID (e.g., "000001" or "BILL-2026-001").
+        """
+        book = get_book()
+        result = book.delete_bill(bill_id=bill_id)
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="delete", entity_type="customer")
+    def delete_customer(customer_id: str) -> str:
+        """Delete a customer with no invoices.
+
+        Customers with any invoices (posted or unposted) cannot be deleted.
+        Delete the invoices first, then delete the customer.
+
+        Args:
+            customer_id: Customer ID (e.g., "000001").
+        """
+        book = get_book()
+        result = book.delete_customer(customer_id=customer_id)
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="write", operation="delete", entity_type="vendor")
+    def delete_vendor(vendor_id: str) -> str:
+        """Delete a vendor with no bills.
+
+        Vendors with any bills (posted or unposted) cannot be deleted.
+        Delete the bills first, then delete the vendor.
+
+        Args:
+            vendor_id: Vendor ID (e.g., "000001").
+        """
+        book = get_book()
+        result = book.delete_vendor(vendor_id=vendor_id)
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="read")
+    def get_outstanding_invoices(
+        owner_type: str | None = None,
+        customer_id: str | None = None,
+        vendor_id: str | None = None,
+    ) -> str:
+        """Get all posted invoices/bills with outstanding balances.
+
+        Args:
+            owner_type: Filter by "customer" or "vendor". Omit for all.
+            customer_id: Filter by specific customer ID.
+            vendor_id: Filter by specific vendor ID.
+        """
+        book = get_book()
+        result = book.get_outstanding_invoices(
+            owner_type=owner_type,
+            customer_id=customer_id,
+            vendor_id=vendor_id,
+        )
+        return _json(result)
+
+    @mcp.tool()
+    @safe_tool
+    @audit_log(classification="read")
+    def vendor_spending_report(
+        start_date: str,
+        end_date: str,
+        vendor_id: str | None = None,
+    ) -> str:
+        """Get spending breakdown by vendor for a period.
+
+        Analyzes posted vendor bills to show total billed, total paid,
+        and outstanding amounts per vendor.
+
+        Args:
+            start_date: Start of period (YYYY-MM-DD).
+            end_date: End of period (YYYY-MM-DD).
+            vendor_id: Optional filter to a specific vendor.
+        """
+        book = get_book()
+        result = book.vendor_spending_report(
+            start_date=start_date,
+            end_date=end_date,
+            vendor_id=vendor_id,
+        )
+        return _json(result)
